requirements.txt
flask==2.3.3
pillow==10.0.0
numpy==1.24.3
werkzeug==2.3.7
opencv-python==4.8.0.76
gunicorn==21.2.0